+++ modules/CPDoS.py
@@ -20,22 +20,20 @@ def check_CPDoS(url, s, req_main, domain, custom_header, authent):
     print("\033[36m ├ CPDoS analyse\033[0m")
 
     uri = "{}?CPDoS={}".format(url, random.randint(1, 100), random.randint(1, 100))
-
+    headers = {'User-agent': 'Mozilla/5.0 (Windows NT 6.3; WOW64; Trident/7.0; LCJB; rv:11.0) like Gecko'}
     try:
-        req_main = requests.get(url, verify=False, allow_redirects=False, timeout=20, auth=authent)
-    except:
-        pass
+        req_main = s.get(url, headers=headers, verify=False, allow_redirects=False, timeout=15, auth=authent)
+        #print(req_main.content)
 
-    req_len = len(req_main.content)
-    main_status_code = req_main.status_code
+        main_len = len(req_main.content)
+        main_status_code = req_main.status_code
 
-    try:
         HHO(uri, s, main_status_code, authent)
         HMC(uri, s, main_status_code, authent)
         HMO(uri, s, main_status_code, authent)
         HHCN(uri, s, authent)
         HBH(url, s, authent)
-        get_error(uri, s, main_status_code, authent)
+        get_error(uri, s, main_status_code, main_len, authent)
         #waf_rules(url, s, main_status_code, authent)
     except KeyboardInterrupt:
         print(" ! Canceled by keyboard interrupt (Ctrl-C)")